+++ backend/show-info-service/src/main/java/group/corona/ServiceApplication.java
@@ -1,4 +1,4 @@
-package com.test;
+package group.corona;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
