+++ backend/show-info-service/src/main/java/group/corona/mapper/GetInfoMapper.java
@@ -1,16 +1,16 @@
-package com.test.service;
+package group.corona.mapper;
 
-import com.test.domain.ChinaInfo;
-import com.test.domain.ProvinceInfo;
-import com.test.domain.WorldInfo;
+import group.corona.domain.ChinaInfo;
+import group.corona.domain.ProvinceInfo;
+import group.corona.domain.WorldInfo;
 
 import java.util.List;
 
 /**
- * 获取疫情信息的service层接口
+ * 疫情数据的dao层接口
  * @author balance
  */
-public interface GetInfoService {
+public interface GetInfoMapper {
 
     /**
      * 获取中国各省的疫情信息
